requirements.txt
openai-whisper
numpy
openai
python-dotenv
pypdf2
rtf-converter
pycryptodome

